requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pandas
scikit-learn
joblib
pydantic